Problem appears to be with Duration#shiftTo
Added the following tests for the issue:
 - Duration#shiftTo hours with decimal
   test/duration/units.test.js
   npm run jest test/duration/units.test.js

 - Duration#toFormat handles decimal
   test/duration/format.test.js
   npm run jest test/duration/format.test.js

There are a few consol.logs that show were .4 hours is being
converted to 23.999999999999993 minutes in the shiftTo funciton.
There is a for loop in there that iterates over the orderedUnits
(year, month, day, etc.). At the end of that loop the acc contains
{ minutes: 0.9999999999999929 }